requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
Pillow
python-dotenv 